+++ test/e2e/jest-runner-vscode.config.js
@@ -4,12 +4,12 @@ const path = require('path');
 
 /** @type {import('jest-runner-vscode').RunnerOptions} */
 const config = {
-	version: '1.56.2',
+	version: '1.71.2',
 	launchArgs: ['--disable-extensions'],
 	openInFolder: true,
 	workspaceDir: path.join(__dirname, 'workspace/workspace.code-workspace'),
 	extensionDevelopmentPath: path.join(__dirname, '../..'),
-	filterOutput: true,
+	// filterOutput: true,
 };
 
 module.exports = config;