+++ test/integration/jest.config.js
@@ -3,8 +3,7 @@
 /** @type {import('@jest/types').Config.InitialOptions} */
 const config = {
 	testMatch: ['<rootDir>/*/test.[jt]s?(x)'],
-	preset: 'ts-jest',
-	globals: { 'ts-jest': { tsconfig: '<rootDir>/../../tsconfig.test.json' } },
+	transform: { '^.+\\.[jt]sx?$': '@swc/jest' },
 	verbose: true,
 	modulePathIgnorePatterns: [
 		'<rootDir>/../../.vscode-test',