+++ src/utils/documents/__tests__/get-fixes.ts
@@ -22,6 +22,7 @@ describe('getFixes', () => {
 				customSyntax: 'postcss-scss',
 			},
 			fix: false,
+			reportDescriptionlessDisables: false,
 			reportNeedlessDisables: true,
 			reportInvalidScopeDisables: false,
 		};